requirements.txt
PyQt5
Pillow
python-chess
chessidle
